Also let CENTER move to 3 places
This way we can use super+5 to focus a window on the center 33% of the
screen, while using super+4 and super+6 to focus windows on the left and
right. Set the center heights to 100 and the center widths to 33, 50 and
100.

(There still is some incidental unexpected behaviour when moving to
some positions. But it appears better than before when not setting the
maximized (-1) value on X separately from Y.)